feat : comments Iter 1

+++ main-service/src/main/java/ru/practicum/ewm/service/CommentServiceImpl.java
@@ -9,13 +9,17 @@ import ru.practicum.ewm.dto.mapper.CommentMapper;
 import ru.practicum.ewm.exception.CommonException;
 import ru.practicum.ewm.model.Comment;
 import ru.practicum.ewm.model.Event;
+import ru.practicum.ewm.model.Request;
 import ru.practicum.ewm.model.User;
+import ru.practicum.ewm.model.enums.EventState;
+import ru.practicum.ewm.model.enums.RequestStatus;
 import ru.practicum.ewm.repository.CommentRepository;
 import ru.practicum.ewm.repository.EventRepository;
 import ru.practicum.ewm.utility.CheckUtil;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -54,7 +58,21 @@ public class CommentServiceImpl implements CommentService {
     public CommentDto addComment(Long userId, Long eventId, CommentDtoText commentDtoText) {
 
         User user = checkUtil.checkUserId(userId);
-        Event event = checkUtil.checkEventInitiator(userId, eventId);
+        Event event = checkUtil.checkEventId(eventId);
+
+        if (Objects.equals(userId, event.getInitiator().getId())) {
+            throw new CommonException("Initiator can't add, update or delete commentaries");
+        }
+
+        if (!event.getState().equals(EventState.FINISHED)) {
+            throw new CommonException("Event is not finished");
+        }
+
+        Request request = checkUtil.checkRequestEventAndRequester(eventId, userId);
+
+        if (!request.getStatus().equals(RequestStatus.CONFIRMED)) {
+            throw new CommonException("User userId = " + userId + "wasn't confirmed to participate in event eventId = " + eventId);
+        }
 
         Comment comment = Comment.builder()
                 .text(commentDtoText.getText())
@@ -73,8 +91,17 @@ public class CommentServiceImpl implements CommentService {
     public CommentDto updateComment(Long userId, Long eventId, Long comId, CommentDtoText commentDtoText) {
 
         checkUtil.checkUserId(userId);
-        checkUtil.checkEventInitiator(userId, eventId);
+        Event event = checkUtil.checkEventId(eventId);
+
+        if (Objects.equals(userId, event.getInitiator().getId())) {
+            throw new CommonException("Initiator can't add, update or delete commentaries");
+        }
+
         Comment comment = checkUtil.checkCommentId(comId);
+
+        if (!comment.getAuthor().getId().equals(userId)) {
+            throw new CommonException("You can't update not your commentary");
+        }
 
         if (comment.getText().equals(commentDtoText.getText())) {
             throw new CommonException("Nothing to update. All is up do date.");
@@ -93,7 +120,17 @@ public class CommentServiceImpl implements CommentService {
     public void deleteCommentByUser(Long userId, Long eventId, Long comId) {
 
         checkUtil.checkUserId(userId);
-        checkUtil.checkEventInitiator(userId, eventId);
+        Event event = checkUtil.checkEventId(eventId);
+
+        if (Objects.equals(userId, event.getInitiator().getId())) {
+            throw new CommonException("Initiator can't add, update or delete commentaries");
+        }
+
+        Comment comment = checkUtil.checkCommentId(comId);
+
+        if (!comment.getAuthor().getId().equals(userId)) {
+            throw new CommonException("You can't delete not your commentary");
+        }
 
         commentRepository.deleteById(comId);
